% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakes.R
\docType{data}
\name{exampleLake}
\alias{exampleLake}
\title{Lake Data for \code{lakemorpho} examples}
\format{SpatialPolygonDataframe with 115 lakes, each with 13 variables}
\description{
This example data is a SpatialPolygonsDataframe 
of lakes named exampleLakes.  These lakes are originally from the 
\href{http://www.horizon-systems.com/nhdplus/NHDPlusV2_home.php}{NHDPlus V2}
dataset.
}
\keyword{datasets}

