\docType{data}
\name{exampleLakes}
\alias{exampleLakes}
\title{Lake Data for \code{lakemorpho} examples}
\format{SpatialPolygonDataframe with 115 lakes, each with 13 variables}
\description{
  This example data is a SpatialPolygonsDataframe of lakes
  named exampleLakes.  These lakes are originally from the
  \href{http://www.horizon-systems.com/nhdplus/NHDPlusV2_home.php}{NHDPlus
  V2} dataset.
}
\keyword{datasets}

