% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeMeanWidth.R
\name{lakeMeanWidth}
\alias{lakeMeanWidth}
\title{Function to return lake Mean Width}
\usage{
lakeMeanWidth(inLakeMorpho)
}
\arguments{
\item{inLakeMorpho}{An object of \code{\link{lakeMorphoClass}}.  Output of the 
\code{\link{lakeSurroundTopo}} function would be appropriate as input}
}
\value{
Returns a numeric value for the mean width of the lake
}
\description{
Mean lake width is the result of lake surface area divded by the maximum 
length.
}
\examples{
\dontrun{
data(lakes)
lakeMeanWidth(inputLM)
}
}
\references{
Florida LAKEWATCH (2001). A Beginner's guide to water management
            - Lake Morphometry (2nd ed.). Gainesville: Florida LAKEWATCH, 
            Department of Fisheries and Aquatic Sciences.
            \href{http://edis.ifas.ufl.edu/pdffiles/FA/FA08100.pdf}{Link}
#' @import rgeos
}

