% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lakeShorelineLength.R
\name{lakeShorelineLength}
\alias{lakeShorelineLength}
\title{Function to calculate shoreline length}
\usage{
lakeShorelineLength(inLakeMorpho)
}
\arguments{
\item{inLakeMorpho}{An object of \code{\link{lakeMorphoClass}}.  Output of the
       \code{\link{lakeSurroundTopo}} function would be appropriate as input}
}
\value{
Returns a numeric value for the length of the lake shoreline
}
\description{
This function returns the length of the shoreline of the lake and is simply
the perimeter of the input lake polyogn.
}
\examples{
data(lakes)
lakeShorelineLength(inputLM)
}
\references{
Florida LAKEWATCH (2001). A Beginner's guide to water management
            - Lake Morphometry (2nd ed.). Gainesville: Florida LAKEWATCH,
            Department of Fisheries and Aquatic Sciences.
            \href{http://edis.ifas.ufl.edu/pdffiles/FA/FA08100.pdf}{Link}
}

