% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeVolume.R
\name{lakeVolume}
\alias{lakeVolume}
\title{Calculates Lake Volume in R}
\usage{
lakeVolume(
  inLakeMorpho,
  zmax = NULL,
  slope_quant = 0.5,
  correctFactor = 1,
  addBathy = FALSE
)
}
\arguments{
\item{inLakeMorpho}{An object of \code{\link{lakeMorphoClass}}.  Output of the
\code{\link{lakeSurroundTopo}} function would be appropriate as input}

\item{zmax}{Maximum depth of the lake.  If none entered and elevation dataset
is inlcuded in inLakeMorpho, \code{\link{lakeMaxDepth}} is used
to estimate a maximum depth.}

\item{slope_quant}{The slope quantile to use to estimate maximum depth.
Defaults to the median as described in (Hollister et. al, 
2011).}

\item{correctFactor}{This a factor used by \code{\link{lakeMaxDepth}} to
correct the predicted maximum lake depth.  Defaults to 1.}

\item{addBathy}{Logical to include a depth raster on the input 
\code{\link{lakeMorphoClass}} object.  This is labelled as 
'pseudoBathy' in the output.  It is depth estimated using 
the maximum lake depth and maximum distance ratio.  Might be
useful for some applications but shouldn't be considered a 
replacement for a bathymetric survey.}
}
\value{
Returns a numeric value for the total volume of the lake
}
\description{
This function returns lake volume for the input lake.  The volume is
calculated using maximum lake depth and maximum distance ratio to transform
all pixels and thus, distances, to an estimated depth.  These depths are
multiplied by the area of the pixel and summed.
}
\examples{
data(lakes)
lakeVolume(inputLM, addBathy = TRUE)
}
\references{
Hollister, J. W., W.B. Milstead (2010). Using GIS to Estimate
            Lake Volume from Limited Data. Lake and Reservoir Management.
            26(3)194-199.
            \doi{10.1080/07438141.2010.504321}

Florida LAKEWATCH (2001). A Beginner's guide to water management
            - Lake Morphometry (2nd ed.). Gainesville: Florida LAKEWATCH,
            Department of Fisheries and Aquatic Sciences.
            \href{http://edis.ifas.ufl.edu/pdffiles/FA/FA08100.pdf}{Link}
}
