% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.lakemorpho.R
\name{plot.lakeMorpho}
\alias{plot.lakeMorpho}
\title{Default plotting of a lakeMorpho object}
\usage{
\method{plot}{lakeMorpho}(x, dist = FALSE, length = TRUE, width = TRUE,
  fetch = FALSE, ...)
}
\arguments{
\item{x}{input lakeMorpho class to plot}

\item{...}{allows for passing of other plot parameters}
}
\description{
Plots the lakeMorpho class by showing lake, surround topography and in lake distance
}
\examples{
\dontrun{
data(lakes)
plot(inputLM)
}
}

