% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lakes.R
\docType{data}
\name{exampleElev}
\alias{exampleElev}
\title{Elevation Data for \code{lakemorpho} examples}
\format{RasterLayer of 1094 x 1419 and 30 meter resolution}
\description{
This example data is a RasterLayer of a small subset of the National
Elevation Dataset named exampleElev.
}
\keyword{datasets}

