% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lakes.R
\docType{data}
\name{inputLM}
\alias{inputLM}
\title{lakemorpho class data for \code{lakemorpho} examples}
\format{lakemorpho class}
\description{
This example lakemorpho class was generated using lakeSurroundTopo with the
included exampleElev and exampleLake data.
}
\keyword{datasets}

