% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lakeMorphoClass.R
\name{lakeMorphoClass}
\alias{lakeMorphoClass}
\title{Function to create a lakeMorpho class - this is input to all other methods}
\usage{
lakeMorphoClass(inLake, inElev = NULL, inCatch = NULL, inLakeDist = NULL,
  lakeOnEdge = F)
}
\arguments{
\item{inLake}{input lake SpatialPolygons object. Required.}

\item{inElev}{input elevation model raster object}

\item{inCatch}{input catchement SpatialPolygons object, can be buffer
around lake}

\item{inLakeDist}{input euclidean distance raster that measures distance
from shore to any pixel in the lake}

\item{lakeOnEdge}{Boolean indicating if inCatch (or lake Buffer) extends
       beyond extent of elevation data}
}
\value{
Returns an object of class 'lakeMorpho'
}
\description{
This is a helper function that creates a lakeMorpho class object
}
\seealso{
lakeSurroundTopo
}

