\name{plot.lakeMorpho}
\alias{plot.lakeMorpho}
\title{Default plotting of a lakeMorpho object}
\usage{
  plot.lakeMorpho(inLakeMorpho, ...)
}
\arguments{
  \item{inLakeMorpho}{input lakeMorpho class to plot}

  \item{...}{allows for passing of other plot parameters
  (although I am not sure this is true)}
}
\description{
  Plots the lakeMorpho class by showing lake, surround
  topography and in lake distance
}

