% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lakemorpho.R
\name{plot.lakeMorpho}
\alias{plot.lakeMorpho}
\title{Default plotting of a lakeMorpho object}
\usage{
\method{plot}{lakeMorpho}(x, dist = FALSE, length = TRUE, width = TRUE, fetch = FALSE, ...)
}
\arguments{
\item{x}{input lakeMorpho class to plot}

\item{dist}{Boolean to control plotting of in lake distance}

\item{length}{Boolean to control plotting of max lake length line}

\item{width}{Boolean to control plotting of max lake width line}

\item{fetch}{Boolean to control plotting of fetch lines}

\item{...}{allows for passing of other plot parameters}
}
\description{
Plots the lakeMorpho class by showing lake, surround topography and in lake distance
}
\examples{
data(lakes)
plot(inputLM)

}
